1545:moving SearchInput into molecules
