+++ components/index.ts
@@ -1,5 +1,4 @@
 export * from './Blog/Blog';
-export * from './SearchInput/SearchInput';
 export * from './ApiError/ApiError';
 export * from './ImageModal';
 export * from './PasswordValidityLevel/PasswordValidityLevel';
