+++ frontends/landing/header/LandingHeader.tsx
@@ -1,5 +1,4 @@
-import { SearchInput } from 'components';
-import { AuthButton, Logo } from 'components/molecules';
+import { AuthButton, Logo, SearchInput } from 'components/molecules';
 import { MobileHeader, TabletHeader } from 'components/organisms';
 import { useDevice } from 'hooks';
 import { FC, useMemo } from 'react';
